short wa?

+++ problems/poj/x_1837/main_1837.cpp
@@ -1,20 +1,20 @@
 #include <cstdio>
 #define S 4725
-int C, G, w[20], d[20], dp[2][S*2];
+short C, G, w[20], d[20], dp[2][S*2];
 int main()
 {
-    scanf("%d%d", &C, &G);
-    for (int i=1; i<=C; ++i) scanf("%d", &d[i]);
+    scanf("%hd%hd", &C, &G);
+    for (short i=1; i<=C; ++i) scanf("%hd", &d[i]);
     dp[0][S] = 1;
-    for (int i=1; i<=G; ++i) {
-        scanf("%d", &w[i]);
+    for (short i=1; i<=G; ++i) {
+        scanf("%hd", &w[i]);
         //w[i-1] = 0;   // TODO: why does uncommenting this break it?
-        for (int j=-S; j<=S; ++j) {
+        for (short j=-S; j<=S; ++j) {
             dp[i&1][j+S] = 0;
-            for (int k=1; k<=C; ++k)
+            for (short k=1; k<=C; ++k)
                 dp[i&1][j+S] += dp[~i&1][j-d[k]*w[i]+S];
         }
     }
-    printf("%d\n", dp[G&1][S]);
+    printf("%hd\n", dp[G&1][S]);
 }
